allow multiple slaves on the same node
This makes the slaves use file locking to exclusively own their
cache directory. If the directory is already locked, they use
a suffix to differentiate their own. This works around the fact
that run_isolate.py isn't safe to run concurrently on the same
cache directory.

It's wasteful in terms of cache capacity, but this is an easier
fix than fixing isolate itself.