Support for server-side retry of "task groups" via "max_retries" parameter.
When a task definition has the "max_retries" parameter, if it fails, the
task will be retried up to that many times (e.g. max_retries=1 will retry once).

This retry is triggered by the slave hitting a retry_task endpoint on the
master which registers a new task and enqueues it.

User-facing UIs like client.py and the job page have been reworked to
track progress based on "task groups" rather than tasks. As part of this,
client.py now sets a meaningful return code.