Support for archiving artifacts of failed tests based on path globs
Use the glob2 module to for recursive globbing support. The list of globs
is provided as part of the task description, and is handled similarly to the
timeout parameter. It's not persisted in the database, only in the JSON
task description that goes into beanstalk.

Matched artifacts are archived into a zip file which is uploaded to S3.
This S3 key is persisted in the backing database.

As part of this commit, also fix the non-uniqueness of the stdout/stderr
S3 keys. This needs to be a compound key based on the compound key in the
dist_test_tasks schema, or else the tasks will squash each other.