client: use ThreadPool for download instead of multiprocessing
This switches to using threads instead of processes for parallel download,
and also adds a workaround for a weird python behavior where KeyboardInterrupt
(^C) isn't handled during a future join() call. Instead, we have to loop around
a join with a timeout.

Tested manually that ^C now works during the artifact download.