Add slave anti-affinity for retry tasks.
Slaves now keep a cache of seen (job_id, task_id). If a retried task
is found in this cache, the slave will release it back to the beanstalk
queue and sleep. Items expire from the cache if the slave sees the same
retry task 10 times, or based on oldest-insertion if the cache is at capacity.

Tested locally by running some tasks with retries.