client: avoid overly-broad 'except' clauses
This fixes an issue where the retry path for 'watch' as well as a few
other cases were using 'except:' rather than 'except Exception:'. This
meant that it would catch KeyboardInterrupt exceptions (which don't
inherit from Exception) and consider them a retriable error.

The result of this was that sometimes ctrl-C wouldn't actually cause
'watch' to exit, which was rather annoying.

I also added a catch for KeyboardInterrupt outside the watch function to
avoid spewing the exception trace during ^C since this is a fairly
common thing for a user to do.