slave: use isolateserver.py download instead of run_isolated.py
We now use the 'download' tool to download the tree to run, rather
than relying on run_isolated to do it for us. This means that we
are now in charge of executing the test command, which should make
it one step easier to do things like run in docker.

This removes support for the tasks which output their own 'isolate'
files used to archive the result. These are not in use by any active
dist-test users, based on a quick query of the production database.