Add retries on downloading task files
This has caused some flakiness for Kudu recently when transient failures
to download artifacts caused a task to fail. This allows the slave to
retry up to three times to download a task, and if it fails, it puts it
back into the queue for someone else to pick up, rather than marking the
task as failed.