Fix handling of commands with relative path
After the recent change to manually run the command instead of using
run_isolated.py we lost some functionality pertaining to paths.

Apparently grind submits jobs with no 'relative_cwd' and with the command set
to 'run_test.sh'. This caused two problems in our code:

- we were always assuming 'relative_cwd' was set. If it's not set it should
  default to the root of the test bundle.
- we were not handling the case that 'command' was a script in the same working
  directory (eg 'run_test.sh'). We assumed it would always be a longer path
  like 'foo/bar/run_test.sh'.

When I tested the prior change I had used Kudu's dist_test integration which
always sets a relative_cwd and runs a script in a subdirectory, so hadn't
noticed the problems.

This change is already running on the deployed infrastructure and appears to be
processing the job queue effectively.